Item branch for development of Sale-Item relation
Created the Item model class; pretty barebones at the moment, will add
more helpers and whatnot later/as we need them. Changes to the Sale
class to allow it to store Items belonging to it, along with the
relevant information, and some functions to go along with it. Everything
is commented (huzzah!)